+++ src/domain/auth/dtos/oidc.public-key.dto.ts
